% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{process_terminal}
\alias{process_terminal}
\title{Emulate terminal}
\usage{
process_terminal(handle)
}
\arguments{
\item{handle}{Process handle obtained from \code{spawn_process}.}
}
\description{
Emulate terminal
}
\examples{
\dontrun{
is_windows <- function () (tolower(.Platform$OS.type) == "windows")

R_binary <- function () {
  R_exe <- ifelse (is_windows(), "R.exe", "R")
  return(file.path(R.home("bin"), R_exe))
}

handle <- spawn_process(R_binary(), c('--no-save'))
Sys.sleep(5)
if(interactive()) process_terminal(handle)
process_kill(handle)
}
}

