% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signal}
\alias{signal}
\title{A helper function used in vignette.}
\usage{
signal(signal, handler)
}
\arguments{
\item{signal}{Signal number.}

\item{handler}{Either \code{"default"} or \code{"ignore"}.}
}
\description{
A helper function used in vignette.
}

