% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subprocess.R
\name{spawn_process}
\alias{spawn_process}
\title{Start a new child process.}
\usage{
spawn_process(command, arguments = character(), environment = character(),
  workdir = "")
}
\arguments{
\item{command}{Path to the executable.}

\item{arguments}{Optional arguments for the program.}

\item{environment}{Optional environment.}

\item{workdir}{Optional new working directory.}
}
\value{
A process handle.
}
\description{
Start a new child process.
}
\details{
\code{environment} can be passed as a \code{character} vector whose
elements take the form \code{"NAME=VALUE"}, a named \code{character}
vector or a named \code{list}.

\code{workdir} is the path to the directory where the new process is
ought to be started. \code{NULL} and \code{""} mean that working
directory is inherited from the parent.
}

