% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{C_tests_utf8}
\alias{C_tests_utf8}
\title{Run UTF8 tests implemented in C.}
\usage{
C_tests_utf8()
}
\value{
A string \code{"All C tests passed!"} if there are no errors.
}
\description{
If there is no error in those tests a simple string message
is returned. If there is at least one error, another message
is returned.
}

